NEXUS-686: The weekdays numbers were invalid between what nexus was expecting, and what the Calendar object was expecting.  This has been corrected (and unit test created)
git-svn-id: file:///opt/svn/repositories/sonatype.org/nexus/trunk@1453 2aa8b3fc-8ebb-4439-a84f-95066eaea8ab

+++ nexus/nexus-configuration/src/test/java/org/sonatype/scheduling/DefaultTaskConfigManagerTest.java
@@ -202,7 +202,7 @@ public class DefaultTaskConfigManagerTest
         else if ( SCHEDULE_TYPE_WEEKLY.equals( type ) )
         {
             Set<Integer> daysToRun = new HashSet<Integer>();
-            daysToRun.add( new Integer( 0 ) );
+            daysToRun.add( new Integer( 1 ) );
             return new WeeklySchedule( (Date) properties.get( PROPERTY_KEY_START_DATE ), (Date) properties
                 .get( PROPERTY_KEY_END_DATE ), daysToRun );
         }
